DOCKER

+++ Study-System/src/main/resources/db/changelog/db.changelog-1.0.sql
@@ -3,23 +3,23 @@ CREATE SCHEMA study_system;
 
 CREATE TABLE study_system.courses
 (
-    id          SERIAL PRIMARY KEY,
+    id          BIGSERIAL PRIMARY KEY,
     name        VARCHAR(32) NOT NULL,
     description TEXT DEFAULT ''
 );
 
 CREATE TABLE study_system.groups
 (
-    id   SERIAL PRIMARY KEY,
+    id   BIGSERIAL PRIMARY KEY,
     name VARCHAR(16) NOT NULL
 );
 
 CREATE TABLE study_system.students
 (
-    id      SERIAL PRIMARY KEY,
+    id      BIGSERIAL PRIMARY KEY,
     name    VARCHAR(32) NOT NULL,
     surname VARCHAR(32) NOT NULL,
-    group_id INT NOT NULL,
+    group_id BIGINT NOT NULL,
     CONSTRAINT fk_students_groups
         FOREIGN KEY (group_id)
         REFERENCES study_system.groups(id)
@@ -29,17 +29,17 @@ CREATE TABLE study_system.students
 
 CREATE TABLE study_system.teachers
 (
-    id      SERIAL PRIMARY KEY,
+    id      BIGSERIAL PRIMARY KEY,
     name    VARCHAR(32) NOT NULL,
     surname VARCHAR(32) NOT NULL
 );
 
 CREATE TABLE study_system.schedules
 (
-    id SERIAL PRIMARY KEY,
-    group_id INT NOT NULL,
-    teacher_id INT NOT NULL,
-    course_id INT NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    group_id BIGINT NOT NULL,
+    teacher_id BIGINT NOT NULL,
+    course_id BIGINT NOT NULL,
     lesson_date DATE NOT NULL,
 
     CONSTRAINT fk_schedules_groups
@@ -60,9 +60,9 @@ ALTER TABLE study_system.schedules ALTER COLUMN lesson_date TYPE TIMESTAMP WITHO
 --changeset bogdan:3
 CREATE TABLE study_system.groups_courses
 (
-    id        SERIAL PRIMARY KEY,
-    group_id  INT NOT NULL,
-    course_id INT NOT NULL,
+    id        BIGSERIAL PRIMARY KEY,
+    group_id  BIGINT NOT NULL,
+    course_id BIGINT NOT NULL,
 
     CONSTRAINT fk_groups_courses
         FOREIGN KEY (group_id)
